chore: split auth-3 test
