+++ packages/amplify-e2e-tests/src/__tests__/auth_3c.test.ts
@@ -18,7 +18,7 @@ const defaultsSettings = {
   name: 'authTest',
 };
 
-describe('amplify add auth...', () => {
+describe('amplify add auth...c', () => {
   let projRoot: string;
   beforeEach(async () => {
     projRoot = await createNewProjectDir('auth');
@@ -27,46 +27,6 @@ describe('amplify add auth...', () => {
   afterEach(async () => {
     await deleteProject(projRoot);
     deleteProjectDir(projRoot);
-  });
-
-  it('...should init a project and add auth with defaults and push, then remove auth and push should clean up trust relationship conditions', async () => {
-    await initJSProjectWithProfile(projRoot, defaultsSettings);
-    await addAuthWithDefault(projRoot, {});
-    await amplifyPushAuth(projRoot);
-
-    const amplifyMeta = getBackendAmplifyMeta(projRoot);
-    const { AuthRoleName, UnauthRoleName } = amplifyMeta.providers.awscloudformation;
-    const cognitoResource = Object.values(amplifyMeta.auth).find((res: any) => res.service === 'Cognito') as any;
-    const idpId = cognitoResource.output.IdentityPoolId;
-
-    expect(AuthRoleName).toHaveValidPolicyConditionMatchingIdpId(idpId);
-    expect(UnauthRoleName).toHaveValidPolicyConditionMatchingIdpId(idpId);
-
-    await removeAuthWithDefault(projRoot);
-    await amplifyPushAuth(projRoot);
-
-    expect(AuthRoleName).not.toHaveValidPolicyConditionMatchingIdpId(idpId);
-    expect(UnauthRoleName).not.toHaveValidPolicyConditionMatchingIdpId(idpId);
-  });
-
-  it('...should init a project with only user pool and no identity pool', async () => {
-    await initJSProjectWithProfile(projRoot, defaultsSettings);
-    await addAuthUserPoolOnly(projRoot, {});
-    await amplifyPushAuth(projRoot);
-    const meta = getProjectMeta(projRoot);
-    const id = Object.keys(meta.auth).map(key => meta.auth[key])[1].output.UserPoolId;
-    const userPool = await getUserPool(id, meta.providers.awscloudformation.Region);
-
-    expect(userPool.UserPool).toBeDefined();
-    const { userPoolGroups } = meta.auth;
-    expect(userPoolGroups.service).toEqual('Cognito-UserPool-Groups');
-    expect(userPoolGroups.providerPlugin).toEqual('awscloudformation');
-    expect(userPoolGroups.dependsOn.length).toBe(1);
-    expect(userPoolGroups.dependsOn[0].category).toBe('auth');
-    expect(userPoolGroups.dependsOn[0].attributes.length).toBe(3);
-    expect(userPoolGroups.dependsOn[0].attributes).toContain('UserPoolId');
-    expect(userPoolGroups.dependsOn[0].attributes).toContain('AppClientIDWeb');
-    expect(userPoolGroups.dependsOn[0].attributes).toContain('AppClientID');
   });
 
   it('...should init a project where all possible options are selected', async () => {
